Fix mode switching due to information loss when casting mode info
SDL3 verifies the requested mode parameters precisely, which can cause the mode switch can fail due to the float to int rounding when converting the SDL3 mode struct to the SDL2 version. Instead of casting directly, use the SDL2 mode parameters to try and find the closest SDL3 display mode.

This fixes video mode switching for modes with refresh rates that are not precisely an integer, and allows for removing the DisplayMode_2to3 function entirely.