Don't set windows as actually modal
Despite the name, SDL_SetWindowModalFor() only set the window parent in SDL2, while making a window modal in SDL3 has additional implications that SDL2 apps may not expect, such as dimming and/or blocking input on the parent until the modal window is dismissed.

Set the parent when this function is called, but otherwise, don't make the window actually modal.